requirements.txt
accelerate==1.7.0
codetiming
datasets>=4.0.0
liger-kernel==0.5.10
mathruler
numpy
omegaconf
pandas
peft==0.16.0
pillow
pyarrow>=15.0.0
pylatexenc
qwen-vl-utils
scipy
matplotlib
ipykernel
sam2
pycocotools
scikit-image
ray==2.46.0
tensordict==0.8.3
safetensors
huggingface-hub
wandb
